// src/app/modules/setup/models/crud/all-setup.ts
export class AllCommonSetup {
    COMMON_ID: number
    TABLE_ID: number
    ORAGANIZATION_CODE: string
    ORG_NAME: string
    DESCRIPTION: string
    VARCHAR_1: string
    VARCHAR_2: any
    VARCHAR_3: any
    VARCHAR_4: any
    VARCHAR_5: any
    VARCHAR_6: any
    VARCHAR_7: any
    CLASS_NAME: string
    NUMBER_1: number
    NUMBER_2: number
    NUMBER_3: number
    NUMBER_4: number
    NUMBER_5: number
    NUMBER_6: number
    NUMBER_7: number
    DATE_1: Date
    DATE_2: Date
    DATE_3: Date
    DATE_4: Date
    DATE_5: Date
    DATE_6: Date
    RowStatus: number
    User_Name: string
}


export class AllSetupNew {
  ORG_CODE: string
  USER_CODE: string
  SET_UP_ID: number
  VALUE_1: string
  VALUE_2: any
  VALUE_3: any
  VALUE_4: any
  VALUE_5: any
  VALUE_6: any
  VALUE_7: any
  VALUE_8: any
  VALUE_9: any
  VALUE_10: any
  VALUE_KEY: any
  ROW_STATUS: number
  ID: number
}