// src/app/modules/setup/models/crud/room-info.ts
export class RoomInfo{
    ORG_CODE: string;
    ROOM_NAME: any;
    CAPACITY: any;
    NUM_OF_ROWS: any;
    NUM_OF_COLUMNS: any;
    USER_CODE: string;
    TEACHER_CODE: string;
    ROOM_ID: number;
    BUILDING_ID: number;
    RowStatus:number;
}

export class RoomList{
    ROOM_ID: number;
    ROOM_NAME: String;  
}

export class BuildingInfo{
    ORG_CODE: string;
    CAMPUS_CODE: string;
    BUILDING_ID: number;
    BUILDING_NAME: any;
    USER_CODE: string;
    RowStatus:number;
}