// src/app/modules/setup/models/crud/holidays-entry.ts
export class HolidaysEntry {
        ORG_CODE: string; 
        HOLIDAY_TYPE: any;
        //HOLIDAY_DATE: Date[];  
        SESSION_CODE: any;
        HOLIDAY_DATE: any[];
}


export class HolidaysEntryVM {
        ORG_CODE: string; 
        HOLIDAY_TYPE: any[];
        HOLIDAY_DATE: any[];  
        SESSION_CODE: any[];
   
}